refactor: remove untested async columnar shuffle
The async shuffle path (spark.comet.columnar.shuffle.async.enabled) has been
disabled by default and lacked meaningful test coverage. Remove it along with
its supporting classes and configs:

- Delete CometShuffleExternalSorterAsync and ShuffleThreadPool
- Merge CometShuffleExternalSorterSync back into CometShuffleExternalSorter
  as a single concrete class (was an interface + factory)
- Strip async branches from CometBypassMergeSortShuffleWriter and
  CometDiskBlockWriter
- Delete spark.comet.columnar.shuffle.async.{enabled,thread.num} configs
- Rename spark.comet.columnar.shuffle.async.max.thread.num to
  spark.comet.columnar.shuffle.max.writers.per.executor (the cap is actually
  on hash-shuffle writers, not threads); keep the old key as an alias via
  withAlternative
- Remove CometAsyncShuffleSuite, DisableAQECometAsyncShuffleSuite, and the
  asyncShuffleEnable machinery from CometColumnarShuffleSuite; drop the
  async benchmark case
- Drop the async config table entry from the JVM shuffle contributor doc

Net: -773 lines.
