+++ spark/src/test/scala/org/apache/spark/sql/benchmark/CometExecBenchmark.scala
@@ -46,7 +46,6 @@ object CometExecBenchmark extends CometBenchmarkBase {
       .set(
         "spark.shuffle.manager",
         "org.apache.spark.sql.comet.execution.shuffle.CometShuffleManager")
-      .set("spark.comet.columnar.shuffle.async.thread.num", "7")
       .set("spark.comet.columnar.shuffle.spill.threshold", "30000")
 
     val sparkSession = SparkSession.builder
