+++ spark/src/test/scala/org/apache/spark/sql/benchmark/CometShuffleBenchmark.scala
@@ -53,7 +53,6 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
       .set(
         "spark.shuffle.manager",
         "org.apache.spark.sql.comet.execution.shuffle.CometShuffleManager")
-      .set("spark.comet.columnar.shuffle.async.thread.num", "7")
       .set("spark.comet.columnar.shuffle.spill.threshold", "30000")
 
     val sparkSession = SparkSession.builder
@@ -109,8 +108,7 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
             CometConf.COMET_EXEC_ENABLED.key -> "true",
             CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
             CometConf.COMET_SHUFFLE_PREFER_DICTIONARY_RATIO.key -> "1.0",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "false") {
+            CometConf.COMET_SHUFFLE_MODE.key -> "jvm") {
             spark
               .sql(
                 s"SELECT ARRAY_REPEAT(CAST(1 AS ${dataType.sql}), 10) AS c1 FROM parquetV1Table")
@@ -166,8 +164,7 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
             CometConf.COMET_EXEC_ENABLED.key -> "true",
             CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
             CometConf.COMET_SHUFFLE_PREFER_DICTIONARY_RATIO.key -> "1.0",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "false") {
+            CometConf.COMET_SHUFFLE_MODE.key -> "jvm") {
             spark
               .sql(
                 s"SELECT STRUCT(CAST(c1 AS ${dataType.sql})," +
@@ -221,8 +218,7 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
             CometConf.COMET_EXEC_ENABLED.key -> "true",
             CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
             CometConf.COMET_SHUFFLE_PREFER_DICTIONARY_RATIO.key -> "1.0",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "false") {
+            CometConf.COMET_SHUFFLE_MODE.key -> "jvm") {
             spark
               .sql("select c1 from parquetV1Table")
               .repartition(partitionNum, Column("c1"))
@@ -236,8 +232,7 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
             CometConf.COMET_EXEC_ENABLED.key -> "true",
             CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
             CometConf.COMET_SHUFFLE_PREFER_DICTIONARY_RATIO.key -> "2.0",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "false") {
+            CometConf.COMET_SHUFFLE_MODE.key -> "jvm") {
             spark
               .sql("select c1 from parquetV1Table")
               .repartition(partitionNum, Column("c1"))
@@ -251,8 +246,7 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
             CometConf.COMET_EXEC_ENABLED.key -> "true",
             CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
             CometConf.COMET_SHUFFLE_PREFER_DICTIONARY_RATIO.key -> "1000000000.0",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "false") {
+            CometConf.COMET_SHUFFLE_MODE.key -> "jvm") {
             spark
               .sql("select c1 from parquetV1Table")
               .repartition(partitionNum, Column("c1"))
@@ -316,22 +310,7 @@ object CometShuffleBenchmark extends CometBenchmarkBase {
             CometConf.COMET_ENABLED.key -> "true",
             CometConf.COMET_EXEC_ENABLED.key -> "true",
             CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "false") {
-            spark
-              .sql("select c1 from parquetV1Table")
-              .repartition(partitionNum, Column("c1"))
-              .noop()
-          }
-        }
-
-        benchmark.addCase("SQL Parquet - Comet (Comet Async JVM Shuffle)") { _ =>
-          withSQLConf(
-            CometConf.COMET_ENABLED.key -> "true",
-            CometConf.COMET_EXEC_ENABLED.key -> "true",
-            CometConf.COMET_EXEC_SHUFFLE_ENABLED.key -> "true",
-            CometConf.COMET_SHUFFLE_MODE.key -> "jvm",
-            CometConf.COMET_COLUMNAR_SHUFFLE_ASYNC_ENABLED.key -> "true") {
+            CometConf.COMET_SHUFFLE_MODE.key -> "jvm") {
             spark
               .sql("select c1 from parquetV1Table")
               .repartition(partitionNum, Column("c1"))
